Viides iteraatio
Muut stat system elementit toimii patisi XP

+++ Assets/Scripts/Stats/Stat.cs
@@ -6,9 +6,9 @@ using UnityEngine;
 public class Stat
 {
     [SerializeField]
-    public float baseValue;
+    public double baseValue;
 
-    public float GetValue()
+    public double GetValue()
     {
         return baseValue;
     }
